+++ CHANGELOG.md
@@ -2,6 +2,7 @@ InfrastructureModels.jl Change Log
 ==================================
 
 ### Staged
+- Added compare_dict function
 - Added arrays_to_dicts! function
 - Fixed bug when matlab function returns a value that is not called "mpc"